pyproject.toml
[project]
name = "solaredge"
version = "0.1.0"
description = "Fetch SolarEdge 15-minute and daily energy data."
requires-python = ">=3.11"
dependencies = [
    "requests>=2.0",
    "python-dotenv>=1.0",
    "playwright>=1.41",
]
